+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/model/TissueList.java
@@ -78,7 +78,7 @@ public class TissueList {
 
     public static List<TissueList> getAllTissueListsForRealm(String realm, String query) {
         List<TissueList> results = new ArrayList<>();
-        HashMap<Long, Tissue> tissues = new HashMap<>();
+        HashMap<Integer, Tissue> tissues = new HashMap<>();
         HashMap<Integer, OncHistoryDetail> oncHistoryDetailHashMap = new HashMap<>();
         HashMap<String, Participant> participantHashMap = new HashMap<>();
         SimpleResult result = inTransaction((conn) -> {
@@ -118,7 +118,7 @@ public class TissueList {
                         }
                     }
                     for (Tissue tissue : tissues.values()) {
-                        Long tissueOncHistoryDetailId = tissue.getOncHistoryDetailId();
+                        Integer tissueOncHistoryDetailId = tissue.getOncHistoryDetailId();
                         OncHistoryDetail oncHistoryDetail = oncHistoryDetailHashMap.get(tissueOncHistoryDetailId);
                         oncHistoryDetail.getTissues().add(tissue);
                     } //  add onchistories to their particiapnt
